pyproject.toml
[tool.poetry]
name = "python-personal-cloud-ftp"
version = "0.1.0"
description = ""
authors = ["ImMamey <32584037+ImMamey@users.noreply.github.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
pyqt5 = "^5.15.7"
pyftpdlib = "^1.5.7"
pydantic = "^1.10.4"
python-dotenv = "^0.21.1"
pywin32 = "^305"
black = "^22.12.0"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
mypy = "^0.991"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
